14enWSL/pyproject.toml
# ═══════════════════════════════════════════════════════════════════════════════
# pyproject.toml — Modern Python Project Configuration
# ═══════════════════════════════════════════════════════════════════════════════
# NETWORKING class — ASE, CSIE | Computer Networks Laboratory
# by ing. dr. Antonio Clim
#
# Week 14: Integrated Recap and Project Evaluation
# ═══════════════════════════════════════════════════════════════════════════════

[project]
name = "netenwsl-week14"
version = "1.0.0"
description = "Week 14 Lab Kit for Computer Networks course - Integrated Recap"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Antonio Clim"}
]
maintainers = [
    {name = "Antonio Clim"}
]
keywords = [
    "networking",
    "education",
    "docker",
    "load-balancing",
    "tcp-ip",
    "wireshark"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Internet",
    "Topic :: System :: Networking",
    "Typing :: Typed"
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "docker>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.1.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/antonioclim/netENwsl"
Documentation = "https://github.com/antonioclim/netENwsl/tree/main/docs"
Repository = "https://github.com/antonioclim/netENwsl.git"
Issues = "https://github.com/antonioclim/netENwsl/issues"

[project.scripts]
week14-quiz = "formative.run_quiz:main"
week14-smoke = "tests.smoke_test:main"

# ═══════════════════════════════════════════════════════════════════════════════
# RUFF CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.ruff]
target-version = "py310"
line-length = 100
indent-width = 4

# Exclude common directories
exclude = [
    ".git",
    ".github",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "artifacts",
    "pcap",
    "docker/volumes",
]

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]

# Ignore these rules
ignore = [
    "E501",   # line too long (handled by formatter)
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring
    "D213",   # Multi-line docstring summary should start at the second line
    "PLR0913", # Too many arguments
    "PLR2004", # Magic value used in comparison
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Tests can use assert and don't need docstrings
"tests/**/*.py" = ["S101", "D103"]
# Scripts are standalone
"scripts/**/*.py" = ["D103"]
# Exercises are for learning
"src/exercises/**/*.py" = ["D103"]
# Homework files are student templates
"homework/**/*.py" = ["D103", "ERA001"]

[tool.ruff.lint.isort]
known-first-party = ["src", "scripts", "formative"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ═══════════════════════════════════════════════════════════════════════════════
# PYTEST CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "docker: marks tests that require Docker",
    "network: marks tests that require network access",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ═══════════════════════════════════════════════════════════════════════════════
# COVERAGE CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.coverage.run]
source = ["src", "scripts", "formative"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# ═══════════════════════════════════════════════════════════════════════════════
# MYPY CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
